Remove mock data, add doc strings, comments and class methods
Removed the mock data stored in self.userdb since the form of
data stored is known. Added doc strings and comments to the code
to explain what the code does and what is expected. Added class
methods to the User class that will be used in the WeConnect class.
Also added print statements to the code for easier troubleshooting.
These will be removed once the code passes tests.